Check if paru installed successfully

+++ install.sh
@@ -1,22 +1,83 @@
 #!/bin/bash
 
+# Reset
+Color_Off='\033[0m'       # Text Reset
+
+# Regular Colors
+Black='\033[0;30m'        # Black
+Red='\033[0;31m'          # Red
+Green='\033[0;32m'        # Green
+Yellow='\033[0;33m'       # Yellow
+Blue='\033[0;34m'         # Blue
+Purple='\033[0;35m'       # Purple
+Cyan='\033[0;36m'         # Cyan
+White='\033[0;37m'        # White
+
+# Bold
+BBlack='\033[1;30m'       # Black
+BRed='\033[1;31m'         # Red
+BGreen='\033[1;32m'       # Green
+BYellow='\033[1;33m'      # Yellow
+BBlue='\033[1;34m'        # Blue
+BPurple='\033[1;35m'      # Purple
+BCyan='\033[1;36m'        # Cyan
+BWhite='\033[1;37m'       # White
+
+# Underline
+UBlack='\033[4;30m'       # Black
+URed='\033[4;31m'         # Red
+UGreen='\033[4;32m'       # Green
+UYellow='\033[4;33m'      # Yellow
+UBlue='\033[4;34m'        # Blue
+UPurple='\033[4;35m'      # Purple
+UCyan='\033[4;36m'        # Cyan
+UWhite='\033[4;37m'       # White
+
+# Background
+On_Black='\033[40m'       # Black
+On_Red='\033[41m'         # Red
+On_Green='\033[42m'       # Green
+On_Yellow='\033[43m'      # Yellow
+On_Blue='\033[44m'        # Blue
+On_Purple='\033[45m'      # Purple
+On_Cyan='\033[46m'        # Cyan
+On_White='\033[47m'       # White
+
+# High Intensity
+IBlack='\033[0;90m'       # Black
+IRed='\033[0;91m'         # Red
+IGreen='\033[0;92m'       # Green
+IYellow='\033[0;93m'      # Yellow
+IBlue='\033[0;94m'        # Blue
+IPurple='\033[0;95m'      # Purple
+ICyan='\033[0;96m'        # Cyan
+IWhite='\033[0;97m'       # White
+
+# Bold High Intensity
+BIBlack='\033[1;90m'      # Black
+BIRed='\033[1;91m'        # Red
+BIGreen='\033[1;92m'      # Green
+BIYellow='\033[1;93m'     # Yellow
+BIBlue='\033[1;94m'       # Blue
+BIPurple='\033[1;95m'     # Purple
+BICyan='\033[1;96m'       # Cyan
+BIWhite='\033[1;97m'      # White
+
+# High Intensity backgrounds
+On_IBlack='\033[0;100m'   # Black
+On_IRed='\033[0;101m'     # Red
+On_IGreen='\033[0;102m'   # Green
+On_IYellow='\033[0;103m'  # Yellow
+On_IBlue='\033[0;104m'    # Blue
+On_IPurple='\033[0;105m'  # Purple
+On_ICyan='\033[0;106m'    # Cyan
+On_IWhite='\033[0;107m'   # White
+
 # Ensure script is not run as root
 if [ "$EUID" -eq 0 ]; then
     echo "Do not run this script as root. Run as a normal user."
     exit 1
 fi
-
-# Vesktop installation prompt
-shopt -s nocasematch
-read -p "Do you want to install Vesktop? (Y/n): " response
-
-if [[ "$response" =~ ^(yes|y)$ ]]; then
-    vesktop_toggle = true
-else
-    vesktop_toggle = false
-fi
-
-shopt -u nocasematch
 
 # System packages
 sudo pacman -S --noconfirm git
@@ -26,6 +87,10 @@ sudo pacman -S --noconfirm --needed base-devel
 git clone https://aur.archlinux.org/paru.git "$HOME/paru"
 cd "$HOME/paru"
 makepkg -si --noconfirm
+if [[ $? -ne 0 ]] ; then
+    printf "${BRed}ERROR WHILE INSTALLING PARU${Color_Off}"
+    exit 1
+fi
 cd ..
 rm -rf "$HOME/paru"
 
@@ -38,7 +103,7 @@ if [sudo lspci | grep -i nvidia > /dev/null]; then
 
     # Prompt for driver choice
     echo "Which NVIDIA driver would you like to use?"
-    echo "1) nvidia (proprietary driver - best performance and recommended) [DEFAULT]"
+    echo -e "1) nvidia (proprietary driver - best performance and recommended) ${BGreen}[DEFAULT]${Color_Off}"
     echo "2) nvidia-open (open source version of the proprietary driver)"
     echo "3) nouveau (fully open source driver - may have limited performance)"
 
@@ -67,8 +132,8 @@ if [sudo lspci | grep -i nvidia > /dev/null]; then
     echo ""
     echo "After installation, reboot your system for changes to take effect."
 else
-    echo "No NVIDIA graphics card was detected on this system.(Its good)"
-    echo "      If you have NVIDIA graphics and still see this message please submit an issue on github."
+    echo -e "${Blue}No NVIDIA graphics card was detected on this system.${BGreen}(Its good)${Color_Off}"
+    echo -e "      ${Red}If you have NVIDIA graphics and still see this message please submit an issue on github.${Color_Off}"
     echo "Your graphics hardware:"
     lspci | grep -E 'VGA|3D|Display' | sed 's/^[^:]*: //'
     sleep 5
